+++ CHANGES.rst
@@ -60,6 +60,7 @@ CHANGES:
 
 FIXED:
 
+  * pull  #476: scenario.status when scenario without steps is skipped (provided by: ar45, jenisys)
   * issue #416: JUnit report messages cut off (submitted by: remcowesterhoud, provided by: bittner)
   * issue #414: Support for Jython 2.7 (submitted by: gabtwi...)
   * issue #384: Active Tags fail with ScenarioOutline (submitted by: BRevzin)